requirements.txt
openai
python-dotenv
faiss-cpu
streamlit